requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
reportlab==4.0.7
Pillow==10.1.0
